Resource Hopper Project- CS633
As a term project for CS633, we designed a perfectly functional application hosted at AWS. The so-called Resource Hoper assembles a balanced software team based on complementary skills of team members. A Manager is able to play what-if scenarios by selecting potential roles for projects. The app was built using Django, Python, Bootstrap, and HTTP.